Introduce IController interface
