+++ src/main/java/HttpServer/Server.java
@@ -1,5 +1,7 @@
 package HttpServer;
 
+import HttpServer.controller.CobSpecController;
+import HttpServer.controller.IController;
 import HttpServer.utility.CommandLineParser;
 import HttpServer.utility.Logger;
 import HttpServer.utility.VerboseLogger;
@@ -14,8 +16,11 @@ public class Server {
         port = commandLineParser.getPort();
         directory = commandLineParser.getDirectory();
 
+        // TODO: Split `core` server (library) from `cob_spec` server (for acceptance testing)
+        IController controller = new CobSpecController();
+
         // Listening. Opens a server socket, accepts a connection, reads/writes, closes the connection. Repeat.
-        SynchronousListener listener = new SynchronousListener(port, directory, logger);
+        SynchronousListener listener = new SynchronousListener(port, directory, controller, logger);
         listener.start();
     }
