+++ src/main/java/HttpServer/controller/IController.java
@@ -1,0 +1,7 @@
+package HttpServer.controller;
+
+import HttpServer.router.Router;
+
+public interface IController {
+    Router getRouter();
+}
